refactor(client): socket chat room 제거
