+++ packages/client/src/App.tsx
@@ -20,15 +20,13 @@ import SearchPage from '~/pages/SearchPage';
 import SettingPage from '~/pages/SettingPage';
 import PostPage from '~/pages/PostPage';
 import WritePage from '~/pages/WritePage';
-import RoomPage from '~/pages/RoomPage';
-import ChatPage from '~/pages/ChatPage';
 import NotFoundPage from '~/pages/NotFoundPage';
 import UserPage, { loader as userPageLoader } from './pages/user/UserPage';
 import UserPostsTab from './pages/user/tabs/UserPostsTab';
 import UserAboutTab from './pages/user/tabs/UserAboutTab';
 import UserSeriesTab from './pages/user/tabs/UserSeriesTab';
 import SeriesPage from './pages/user/SeriesPage';
-import UserErrorBoundary from './pages/UserErrorBoundary';
+import UserErrorBoundary from './pages/user/UserErrorBoundary';
 
 const App = () => {
   useGetMe();
@@ -38,11 +36,10 @@ const App = () => {
   const router = createBrowserRouter([
     { path: '/', element: <HomePage /> },
     { path: '/write', element: <WritePage /> },
-    { path: '/room', element: <RoomPage /> },
     { path: '/search', element: <SearchPage /> },
     { path: '/setting', element: <SettingPage /> },
     {
-      path: '/:_username',
+      path: '/user/:username',
       element: <UserPage />,
       loader: userPageLoader(queryClient),
       errorElement: <UserErrorBoundary />,
@@ -52,9 +49,8 @@ const App = () => {
         { path: 'series', element: <UserSeriesTab /> },
       ],
     },
-    { path: '/:_username/series/:name', element: <SeriesPage /> },
-    { path: '/post/:slug', element: <PostPage /> },
-    { path: '/room/:roomId', element: <ChatPage /> },
+    { path: '/user/:username/series/:seriesName', element: <SeriesPage /> },
+    { path: '/user/:username/post/:slug', element: <PostPage /> },
     { path: '*', element: <NotFoundPage /> },
   ]);
