--- packages/client/src/hooks/queries/room/index.ts
@@ -1,4 +1,0 @@
-export { default as useCreateRoom } from './useCreateRoom';
-export { default as useGetRoom } from './useGetRoom';
-export { default as useGetRooms } from './useGetRooms';
-export { default as useDeleteRoom } from './useDeleteRoom';
