--- packages/client/src/hooks/queries/room/useGetRooms.ts
@@ -1,17 +1,0 @@
-import RoomAPI from '~/lib/api/room';
-import { useQuery } from '@tanstack/react-query';
-import type { UseQueryOptionsOf } from '~/hooks/queries/types';
-
-const useGetRooms = (
-  options: UseQueryOptionsOf<typeof RoomAPI.getRooms> = {},
-) => {
-  return useQuery(getKey(), fetcher(), options);
-};
-
-const getKey = () => ['useGetRooms'];
-const fetcher = () => () => RoomAPI.getRooms();
-
-useGetRooms.getKey = getKey;
-useGetRooms.fetcher = fetcher;
-
-export default useGetRooms;
